ultimas peticiones, display de video y usuario, lista de reproduccion de video y usuario (con Jose)

+++ src/main/java/com/unicast/unicast_backend/controllers/ReproductionListController.java
@@ -75,6 +75,11 @@ public class ReproductionListController {
             throw new Error();
         }
 
+        if (reproList.getName().equals("Favoritos")) {
+            // TODO: gestionar excepcion en condiciones
+            throw new Error();
+        }
+
         reproductionListRepository.delete(reproList);
 
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
